working basic bundling
