+++ new-playground/tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "module": "esnext",
     "target": "es5",
-    "sourceMap": true
+    "sourceMap": true,
+    "skipLibCheck": true
   },
   "exclude": [
     "node_modules"
