+++ package.json
@@ -34,6 +34,8 @@
     "autoprefixer": "^9.6.1",
     "bundle-stats": "^1.10.1",
     "chalk": "2.4.1",
+    "css-loader": "3.2.0",
+    "file-loader": "4.2.0",
     "fork-ts-checker-notifier-webpack-plugin": "^1.0.2",
     "fork-ts-checker-webpack-plugin": "^1.5.0",
     "html-loader": "^0.5.5",
